+++ eslint.config.js
@@ -1,0 +1,50 @@
+import { defineConfig } from "eslint/config";
+import globals from "globals";
+import js from "@eslint/js";
+import prettier from "eslint-plugin-prettier";
+import babelParser from "@babel/eslint-parser";
+
+export default defineConfig([
+  {
+    files: ["**/*.{js,mjs,cjs}"],
+    plugins: {
+      prettier,
+    },
+    languageOptions: {
+      parser: babelParser,
+      parserOptions: {
+        ecmaVersion: 2023,
+        sourceType: "module",
+        requireConfigFile: false
+      },
+    },
+    rules: {
+      "no-unused-vars": [
+        "warn",
+        {
+          argsIgnorePattern: "^_",
+          varsIgnorePattern: "^_",
+          caughtErrorsIgnorePattern: "^_",
+        },
+      ],
+      /*
+      "prettier/prettier": [
+        "info",
+        {
+          tabWidth: 2,
+          useTabs: false,
+        },
+      ],
+      */
+    },
+  },
+  {
+    files: ["**/*.{js,mjs,cjs}"],
+    languageOptions: { globals: globals.browser },
+  },
+  {
+    files: ["**/*.{js,mjs,cjs}"],
+    plugins: { js },
+    extends: ["js/recommended"],
+  },
+]);