+++ index.js
@@ -1,14 +1,14 @@
 /**
- * v6
+ * v7
  *
  * @url https://github.com/homebridge-plugins/homebridge-tado
- * @author SeydX <seyd55@outlook.de>
+ * @author maintained by smart7324 <34960772+smart7324@users.noreply.github.com> - author SeydX <seyd55@outlook.de>
  *
  **/
 
-'use strict';
+import TadoPlatformModule from './src/platform.js';
 
-module.exports = function (homebridge) {
-  let TadoPlatform = require('./src/platform.js')(homebridge);
+export default function (homebridge) {
+  let TadoPlatform = TadoPlatformModule(homebridge);
   homebridge.registerPlatform('@homebridge-plugins/homebridge-tado', 'TadoPlatform', TadoPlatform, true);
 };