+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # Changelog
 
+## v7.5.1 - 2025-03-26
+- Fixed package.json required node version
+
+## v7.5.0 - 2025-03-25
+- Converted package to ES module
+- Activated eslint
+- Updated dependencies
+- Added account name to authentication url message
+- Dropped support for node 18 due to got
+
+## v7.4.3 - 2025-03-19
+- Updated fakegato-history for Homebridge v2.0 support
+- Removed workaround for missing perms enum values
+
+## v7.4.2 - 2025-03-16
+- Updated config.schema.json
+
+## v7.4.1 - 2025-03-14
+- Updated changelog
+
+## v7.4.0 - 2025-03-14
+- Implemented the new authentication workflow for the tado° REST API based on their [official instructions](https://support.tado.com/en/articles/8565472-how-do-i-authenticate-to-access-the-rest-api)
+- Added full support for Homebridge v2.0 (it is also shown as compatible when using Homebridge v1.X)
+- Updated all dependencies to their latest version
+- Added new authentication workflow to config-ui
+- Fixed an issue that caused HomeKit from pairing with the bridge (Error: Accessory out of compliance)
+- Fixed an issue where the plugin crashed after startup due to a type error after upgrading got
+- Removed AirQuality feature [homebridge-tado-platform/issues/152#issuecomment-2708942491](https://github.com/seydx/homebridge-tado-platform/issues/152#issuecomment-2708942491)
+- Fixed an issue where the current temperature has been set to the target temperature after changing the target temperature through Apple Home
+- Fixed an issue that caused Apple Home from pairing with the bridge (Error: Accessory out of compliance)
+- Fixed an issue where the plugin crashed after startup due to errors after upgrading got
+- Fixed minor bugs
+
 ## v6.0.14 - 2021-05-19
 - Fixed an issue where the thermostat displayed wrong room temperature due to wrong celsius/fahrenheit calculation
 - Bump dependencies